Update Schedule.jsx
React.memo is a higher-order component that is used to optimize the performance of React functional components. It is similar to React.PureComponent for class components.

When a component is wrapped with React.memo, it will only re-render if its props have changed. If the props have not changed, the component will not re-render, and instead, the previous result will be used