Worked on Prize Page
Removed some the files that are no longer needed, also merged my work with my teammate's. Resolved conflicts in designs: there were some bugs when doing so. The general layout is done now. Still have many things to work on. Added best hack and second best hack onto the page. Also added a image of trophy, which I photoshop and will most likely to be changed until I find a better image.  Also worked on the styling and layout of the prize page to make it consistent throughout.