Add more content and re-styled prize page
I organized the files, put all the image under the same directories. Fixed the color of the animated header that advocate for the total prize: used a color picker and tested various combinations. Changed the color of gear using photoshop app: which is still subject to change, when a better image, or combination is found. Changed the trophy to another image, and have the award title and prize amount show on it.