Edited Prize Page
I changed the styling for the top prize so that the actual award itself stands out than the money. I also adjusted the page for smaller screen size including a mobile phone. Which the items of the page will get relatively smaller. Also changed the transitions of prize amount, added animations so that it differ from the award title. Also changed the light for podium.