Edited Other Prizes
Fixed it so that the overflow contents will show in second line for smaller display. Also edited the tabs so that the background changes when it is hover, so it is apparent that user can click it. Also edited the animations so that the gears start rotating when hover, instead merely after clicked.