Fixed the code and styling
I recode the page so that the components for the pop ups are ready to be imported in a separate jsx file for modularity. I also updated the styling so that it looks much more appealing and easier for user to navigate across. Still need some minor adjustment such as presentation on smaller screen size.